🎨 refactor : User Model 공통 사용 변경
