+++ lib/blocs/sign_up/sign_up_state.dart
@@ -1,5 +1,5 @@
 import 'package:cakeke/data/models/sign_up/service_purpose.dart';
-import 'package:cakeke/data/models/sign_up/user.dart';
+import 'package:cakeke/data/models/common/user.dart';
 import 'package:equatable/equatable.dart';
 
 class SignUpState extends Equatable {
